[Devendra]Added: CURD operation by using JWT token

+++ src/main/java/com/bridgelabz/AddressBook/repository/AddressBookRepo.java
@@ -6,10 +6,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.bridgelabz.AddressBook.model.AddressBookModel;
 
-public interface AddressBookRepo extends JpaRepository<AddressBookModel, Integer> {
+public interface AddressBookRepo extends JpaRepository<AddressBookModel, Long> {
 
     List<AddressBookModel> findByName(String name);
 
     List<AddressBookModel> findByCity(String city);
-
 }
